pyproject.toml
[project]
name = "sales-portal-tests"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.49",
    "pytest>=8.0",
    "pytest-playwright>=0.6",
    "jsonschema>=4.23",
    "pydantic>=2.10",
    "pytest-check>=2.3",
    "faker>=33.0",
    "pymongo>=4.10",
    "python-dotenv>=1.1",
    "python-dateutil>=2.9",
    "python-telegram-bot>=21.0",
    "allure-pytest>=2.13",
    "allure-python-commons>=2.13",
    "pytest-rerunfailures>=14.0",
    "pytest-timeout>=2.3",
    "pytest-xdist>=3.5",
    "pytest-order>=1.3",
]
# NOTE: No `requests` — Playwright's APIRequestContext handles all HTTP calls.

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "mypy>=1.14",
    "pre-commit>=4.0",
    "pytest-html>=4.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "smoke: critical-path tests",
    "regression: full regression suite",
    "api: API tests",
    "ui: UI tests",
    "products: product-related tests",
    "customers: customer-related tests",
    "orders: order-related tests",
    "managers: manager-related tests",
]
addopts = "--strict-markers --tb=short -v"

# Centralised timeouts/retries defaults (override per-suite via CLI if needed)
timeout = 60
timeout_method = "thread"
reruns = 1
reruns_delay = 2

[tool.mypy]
strict = true
python_version = "3.12"
plugins = ["pydantic.mypy"]
exclude = [".venv"]
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
